refactoring: introduce FilesController
- add new FilesController aimed to incapsulate File Storage Rest API
- move methods related to file management from TalksController to FileController
