+++ conference/webapp/src/main/java/web/controller/TalksController.java
@@ -177,72 +177,4 @@ public class TalksController {
         return ok(messageDto);
     }
 
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping(value = "/talk/{talk_id}/takeFileName",
-            produces = "application/json")
-    public ResponseEntity<Map<String, String>> getFileName(@PathVariable("talk_id") Long talkId) {
-        Talk talk = talkService.findTalkById(talkId);
-
-        File file = storageService.getFile(talk.getPathToAttachedFile());
-        Map<String, String> map = new HashMap<>();
-        map.put("fileName", file.getName());
-        return ok(map);
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping(value = "/talk/{talk_id}/takeFile")
-    public ResponseEntity<InputStreamResource> takeFile(@PathVariable("talk_id") Long talkId) {
-        TalkDto talkDto = talkService.findById(talkId);
-        String filePath = talkService.getFilePath(talkDto);
-
-        File file = storageService.getFile(filePath);
-
-        String mimeType = storageService.getFileTypeIfSupported(file);
-
-        try {
-            InputStreamResource inputStreamResource = new InputStreamResource(new FileInputStream(file));
-            HttpHeaders header = new HttpHeaders();
-            header.setContentType(new MediaType(mimeType.split("/")[0], mimeType.split("/")[1]));
-            header.setContentLength(file.length());
-            header.set("Content-Disposition", "attachment; filename=" + new String(file.getName().getBytes(), StandardCharsets.ISO_8859_1));
-            return new ResponseEntity<>(inputStreamResource, header, HttpStatus.OK);
-        } catch (IOException e) {
-            log.debug(e);
-            return badRequest().build();
-        }
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @PostMapping("/talk/{talk_id}/uploadFile")
-    public ResponseEntity upload(@PathVariable("talk_id") Long talkId,
-                                 @RequestPart(value = "file") MultipartFile file,
-                                 HttpServletRequest request) {
-        String filePath = uploadFile(file);
-        TalkDto talkDto = talkService.findById(talkId);
-        talkService.addFile(talkDto, filePath);
-
-        return ok().build();
-    }
-
-    private String uploadFile(MultipartFile file) {
-        try {
-            return storageService.saveFile(file, FileStorageServiceImpl.FileType.FILE);
-        } catch (IOException e) {
-            log.warn(e);
-            return null;
-        }
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @DeleteMapping("/talk/{talk_id}/deleteFile")
-    public ResponseEntity delete(@PathVariable("talk_id") Long talkId) {
-        TalkDto talkDto = talkService.findById(talkId);
-        String filePath = talkService.getFilePath(talkDto);
-
-        storageService.deleteFile(filePath);
-        talkService.deleteFile(talkDto, true);
-
-        return ok().build();
-    }
-
 }
